ai-worker/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.4
google-cloud-firestore==2.16.0
google-cloud-pubsub==2.21.1
vertexai==1.52.0
elastic-resume-base-hermes @ file://../shared/hermes
